Azure PDF Summarizer/requirements.txt
streamlit
azure-core
azure-ai-formrecognizer
azure-ai-textanalytics
fpdf
fitz
PyMuPDF
